+++ FA2sp/Ext/CMapValidator/Body.h
@@ -5,6 +5,9 @@
 class CMapValidatorExt : public CMapValidator
 {
 public:
+	void ValidateStructureOverlapping(BOOL& result);
+	void ValidateMissingParams(BOOL& result);
+
 	ppmfc::CString FetchLanguageString(const char* Key, const char* def);
 	void InsertString(const char* String, bool IsWarning);
 };