+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Reimplemented file reading system
 - Support Ares Custom Foundation
 - Map validator now checks overlapping structures, raise an error instead of warning for now. String can be modified by `MV_OverlapStructures`
+- Map validator now checks missing logic params, raise an error instead of warning for now. String can be modified by `MV_LogicMissingParams`
 
 ## RELEASE 1.5.2 (2023-03-03)
 - Fixed the bug that money calculation is incorrect